Allow using this plugin in monorepos
The current way of selecting files matches `git diff` results against
the source paths configured in the sfdx config file. Unfortunately, this
prevents having the source code into a sub directory, like in a monorepo
context. In my case, the code is in a `apps/salesforce` directory no
change is detected. The plugin crashed with the following message:

> `WARNING: No changes found!`

This PR is a very raw proposal to shine a light on this issue and start
the discussion about how it should be solved.

The way I do it is a bit hacky and uses a new flag as an example. I have
not tested it yet.

I think a better way would be to use the following command to detect the
relative path of files to the git root.

```
git rev-parse --show-prefix
```

This requires a lot more work but I think it would make the plugin more
versatile.